feat(retrieval): implement GraphRAG hybrid retrieval (ADR-0070)
Implemented a unified GraphRAG retriever that combines vector similarity
(anchors) with graph traversal (expansion) and blended scoring:
score = sim_weight * cosine + graph_weight * (1/(1+hops)) * path_strength.

Changes:
- Implemented core GraphRAG logic in `src/retrieval/graph_rag.rs` using `VecDeque` for BFS
  and tracking cumulative path strength (bottleneck edge weight).
- Integrated `probe_with_graph` and `probe_text_with_graph` via extension module
  `src/framework_graph_rag.rs` to maintain `src/framework.rs` under 500 LOC.
- Exposed GraphRAG methods in WASM via `src/wasm_ext.rs`.
- Added `csm probe-graph` CLI subcommand in `src/cli/commands/probe_graph.rs`.
- Added integration tests in `tests/graph_rag.rs` covering structure,
  cycles, and isolated cases.
- Added performance benchmark `probe_graph_1k_concepts_5_hops` in
  `benches/benchmark.rs` (~0.8ms p50).
- Updated `llms.txt` to reflect new public API surface.

Resolves #140.

Fixed CI failures:
- Resolved linting issues in `tests/graph_rag.rs` (clone-on-copy and float-cmp).
- Fixed LOC gate failure for `src/framework.rs` by using extension module.
