+++ llms-full.txt
@@ -360,6 +360,10 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 - pub enum MemoryEvent
 - impl ChaoticSemanticFramework
 
+### src/framework_graph_rag.rs
+
+- impl ChaoticSemanticFramework
+
 ### src/framework_metrics.rs
 
 - pub struct FrameworkMetrics
@@ -2671,8 +2675,6 @@ impl ChaoticSemanticFramework {
     pub fn incoming_associations(&self, id: &str) -> Result<Vec<(String, f32)>>;
     pub fn shortest_path_hops(&self, from: &str, to: &str) -> Result<Option<Vec<String>>>;
     pub fn get_concept(&self, id: &str) -> Result<Option<Concept>>;
-    pub fn probe_with_graph(&self, query: HVec10240, config: GraphRagConfig) -> Result<Vec<GraphRagResult>>;
-    pub fn probe_text_with_graph(&self, text: &str, config: GraphRagConfig) -> Result<Vec<GraphRagResult>>;
     pub fn persist(&self) -> Result<()>;
     pub fn persistence_health_check(&self) -> Result<()>;
     pub fn load_replace(&self) -> Result<()>;
@@ -2797,6 +2799,17 @@ impl ChaoticSemanticFramework {
     pub fn update_concept_vector(&self, id: &str, vector: HVec10240) -> Result<()>;
     pub fn update_concept_metadata(&self, id: &str, metadata: HashMap<String, serde_json::Value>) -> Result<()>;
     pub fn disassociate(&self, from: &str, to: &str) -> Result<()>;
+}
+```
+
+## src/framework_graph_rag.rs
+
+### impl ChaoticSemanticFramework
+
+```rust
+impl ChaoticSemanticFramework {
+    pub fn probe_with_graph(&self, query: HVec10240, config: GraphRagConfig) -> Result<Vec<GraphRagResult>>;
+    pub fn probe_text_with_graph(&self, text: &str, config: GraphRagConfig) -> Result<Vec<GraphRagResult>>;
 }
 ```
