+++ src/framework.rs
@@ -14,7 +14,6 @@ use crate::metadata_filter::MetadataFilter;
 #[cfg(feature = "persistence")]
 use crate::persistence::Persistence;
 use crate::reservoir::ChaoticReservoir;
-use crate::retrieval::{GraphRagConfig, GraphRagResult, graph_rag_retrieve};
 use crate::singularity::{Concept, ConceptBuilder, Singularity, unix_now_secs};
 
 /// Main framework for chaotic semantic memory
@@ -309,42 +308,6 @@ impl ChaoticSemanticFramework {
         Self::validate_concept_id(id)?;
         let sing = self.singularity.read().await;
         Ok(sing.get(id).cloned())
-    }
-
-    /// GraphRAG retrieval: similarity + graph traversal hybrid.
-    ///
-    /// Combines vector similarity with graph traversal for unified retrieval:
-    /// 1. Anchor: probe(query, anchor_top_k) → seed set
-    /// 2. Expand: traverse from each anchor
-    /// 3. Score: similarity_weight * cosine + graph_weight * (1/(1+hops)) * strength
-    /// 4. Dedupe + rank by score
-    #[instrument(err, skip(self, query, config))]
-    #[allow(clippy::significant_drop_tightening)] // Lock needed for concept and association access
-    pub async fn probe_with_graph(
-        &self,
-        query: HVec10240,
-        config: GraphRagConfig,
-    ) -> Result<Vec<GraphRagResult>> {
-        self.validate_top_k(config.anchor_top_k)?;
-        self.validate_top_k(config.final_top_k)?;
-
-        let sing = self.singularity.read().await;
-        let concepts = sing.all_concepts();
-        let associations = sing.all_associations();
-
-        graph_rag_retrieve(&query, &concepts, &associations, &config)
-    }
-
-    /// GraphRAG retrieval using encoder for text query.
-    #[instrument(err, skip(self, text, config))]
-    pub async fn probe_text_with_graph(
-        &self,
-        text: &str,
-        config: GraphRagConfig,
-    ) -> Result<Vec<GraphRagResult>> {
-        let encoder = crate::encoder::TextEncoder::new();
-        let query = encoder.encode(text);
-        self.probe_with_graph(query, config).await
     }
 
     /// Persist all data to storage
