+++ src/lib.rs
@@ -58,6 +58,7 @@ pub mod framework;
 mod framework_bridge;
 pub mod framework_builder;
 mod framework_events;
+mod framework_graph_rag;
 mod framework_metrics;
 #[cfg(not(target_arch = "wasm32"))]
 mod framework_ops;
